Update entry detail widget on db change
