+++ lotti/lib/widgets/journal/journal_card.dart
@@ -5,6 +5,7 @@ import 'package:lotti/blocs/audio/player_state.dart';
 import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/classes/measurables.dart';
 import 'package:lotti/theme.dart';
+import 'package:lotti/widgets/journal/entry_detail_route.dart';
 import 'package:lotti/widgets/journal/entry_detail_widget.dart';
 import 'package:lotti/widgets/journal/entry_tools.dart';
 import 'package:lotti/widgets/misc/survey_summary.dart';
@@ -170,7 +171,7 @@ class JournalCard extends StatelessWidget {
             Navigator.of(context).push(
               MaterialPageRoute(
                 builder: (BuildContext context) {
-                  return DetailRoute(
+                  return EntryDetailRoute(
                     item: item,
                     index: index,
                   );
@@ -181,44 +182,5 @@ class JournalCard extends StatelessWidget {
         ),
       );
     });
-  }
-}
-
-class DetailRoute extends StatelessWidget {
-  const DetailRoute({
-    Key? key,
-    required this.item,
-    required this.index,
-  }) : super(key: key);
-
-  final int index;
-  final JournalEntity item;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: AppColors.bodyBgColor,
-      appBar: AppBar(
-        title: Text(
-          df.format(item.meta.dateFrom),
-          style: TextStyle(
-            color: AppColors.entryBgColor,
-            fontFamily: 'Oswald',
-          ),
-        ),
-        backgroundColor: AppColors.headerBgColor,
-      ),
-      body: Container(
-        color: AppColors.bodyBgColor,
-        padding: EdgeInsets.only(
-          bottom: MediaQuery.of(context).viewInsets.bottom,
-        ),
-        child: SingleChildScrollView(
-          child: EntryDetailWidget(
-            item: item,
-          ),
-        ),
-      ),
-    );
   }
 }
